move functions into their own locations
